+++ src/main/java/de/jeff_media/InvUnload/Messages.java
@@ -5,15 +5,15 @@ import org.bukkit.ChatColor;
 public class Messages {
 	Main plugin;
 
-	final String PREFIX, MSG_COULD_NOT_UNLOAD, MSG_RADIUS_TOO_HIGH, MSG_NOT_A_NUMBER, MSG_NO_CHESTS_NEARBY;
+	final String PREFIX, MSG_COULD_NOT_UNLOAD, MSG_RADIUS_TOO_HIGH, MSG_NOT_A_NUMBER, MSG_NO_CHESTS_NEARBY, MSG_INVENTORY_EMPTY;
 
 	Messages(Main plugin) {
 		this.plugin = plugin;
 		
-		PREFIX = ChatColor.GOLD + "[InvUnload] "+ChatColor.RESET;
+		PREFIX = ChatColor.translateAlternateColorCodes('&', plugin.getConfig().getString("message-prefix"));
 
 		MSG_COULD_NOT_UNLOAD = PREFIX + ChatColor.translateAlternateColorCodes('&', plugin.getConfig()
-				.getString("message-could-not-unload", "&7There was nothing to unload. Maybe you have no chests for the remaining items?"));
+				.getString("message-could-not-unload", "&7Nothing to unload: There are no chests for the remaining items."));
 		
 		MSG_RADIUS_TOO_HIGH = PREFIX + ChatColor.translateAlternateColorCodes('&', String.format(plugin.getConfig()
 				.getString("message-radius-too-high","&cError:&7 The radius cannot be higher than %d blocks."),plugin.getConfig().getInt("max-chest-radius",20)));
@@ -22,10 +22,10 @@ public class Messages {
 				.getString("message-error-not-a-number", "&cError:&7 '%s' is not a valid number."));
 		
 		MSG_NO_CHESTS_NEARBY = PREFIX+ ChatColor.translateAlternateColorCodes('&', plugin.getConfig()
-				.getString("message-no-chests-nearby","&7There are no chests nearby. Adjust the radius or walk closer to your chests."));
+				.getString("message-no-chests-nearby","&7Nothing to unload: There are no chests nearby. Adjust the radius or walk closer to your chests."));
 		
-		//MSG_INVENTORY_EMPTY = PREFIX+ ChatColor.translateAlternateColorCodes('&', plugin.getConfig()
-		//		.getString("message-inventory-empty","&7Nothing to dump: Your inventory was already empty."));
+		MSG_INVENTORY_EMPTY = PREFIX+ ChatColor.translateAlternateColorCodes('&', plugin.getConfig()
+				.getString("message-inventory-empty","&7Nothing to unload: Your inventory is already empty."));
 		
 		
 	}