+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
-## 3.0
+## 3.0-SNAPSHOT6
 - Rewrote plugin from scratch. Source code is now muuuuch cleaner and faster
 - Unload/dump now works with chests, double chests, shulkerboxes and barrels
+- /dump will no longer take items from the player's hotbar
+- InvUnload can use ChestSort to sort chests that have been affected by unloading/dumping. This can be disabled in the config and only works if the player has automatic chest sorting enabled
+- Particle types and sound effects can be set in the config.yml